move hw access to peripherals 	comment out volume ctl related code 	as this is not working yet
