+++ device/spotify-clock-client/src/main.rs
@@ -1,16 +1,9 @@
 mod spotify_client;
+mod peripherals;
 
 use std::{error::Error, time::Duration};
-use debounce::EventDebouncer;
-use pcf8591::{PCF8591, Pin};
-use tokio::sync::mpsc;
-use futures::executor;
-use librespot::connect::spirc::Spirc;
-use rppal::{
-    gpio::{Gpio, Trigger},
-    system::DeviceInfo,
-};
-use rppal::gpio::Level;
+use rppal::gpio::Gpio;
+use rppal::system::DeviceInfo;
 use tokio::time::sleep;
 
 async fn blink_led() -> Result<(), Box<dyn Error>> {
@@ -27,58 +20,10 @@ async fn blink_led() -> Result<(), Box<dyn Error>> {
     }
 }
 
-async fn set_alsa_volume(volume_percent: u16) {
-    let volume_percent = volume_percent / 5;
-    let mixer = alsa::mixer::Mixer::new("default", false).unwrap();
-    let selem_id = alsa::mixer::SelemId::new("PCM", 0);
-    let selem = mixer.find_selem(&selem_id).unwrap();
-    let (min, max) = selem.get_playback_volume_range();
 
-    let resolution = (max - min) as f64;
-    // let factor: u16 = (((0xFFFF + 1) / resolution) - 1) as u16;
-    let volume: i64 = (volume_percent as f64 / 100.0 * (resolution)) as i64;
-    println!("Setting volume: {:?}, max: {}, §§min: {}", volume, max, min);
-    selem.set_playback_volume_all(volume).unwrap();
-}
-
-async fn read_input(spirc: Spirc) -> Result<(), Box<dyn Error>> {
-    let (sender, mut receiver) = mpsc::channel::<Level>(32);
-
-    // Set up GPI for play/pause button
-    const GPIO_BUTTON: u8 = 24;
-    let mut input = Gpio::new()?.get(GPIO_BUTTON)?.into_input();
-    let callback = move |level| {
-        executor::block_on(sender.send(level)).expect("TODO: panic message");
-    };
-    input.set_async_interrupt(Trigger::RisingEdge, callback).expect("TODO: panic message");
-
-    // Set up i2c for potentiometer ADC
-    let mut converter = PCF8591::new("/dev/i2c-1", 0x48, 5.0).unwrap();
-    let mut current_volume_percent: u16 = 0;
-
-    let delay = Duration::from_millis(100);
-    let debouncer = EventDebouncer::new(delay, move |data: Level| {
-        println!("Play/Pause");
-        spirc.play_pause();
-    });
-    loop {
-        let gpio_level = receiver.recv().await.unwrap();
-        let current_volume_v = converter.analog_read(Pin::AIN0).unwrap();
-        let volume_percent = (current_volume_v / 5.0 * 100.0).round() as u16;
-        if volume_percent != current_volume_percent {
-            set_alsa_volume(volume_percent).await;
-            println!("Set volume {}, previously {}", volume_percent, current_volume_percent);
-            current_volume_percent = volume_percent;
-        }
-
-        if gpio_level == Level::High {
-            debouncer.put(gpio_level);
-        }
-    }
-}
 
 #[tokio::main]
 async fn main() {
-    let (connect_client, connect_client_task) = spotify_client::new().await;
-    let (_first, _second, _third) = tokio::join!(connect_client_task, blink_led(), read_input(connect_client));
+    let (connect_client, connect_client_task) = spotify_client::init().await;
+    let (_first, _second, _third) = tokio::join!(connect_client_task, blink_led(), peripherals::read_input(connect_client));
 }
